adding tests to the client/config package
